Validate fields after creating the runner.
This is the ParentRunner's standard approach for validating.
BlockJUnit4ClassRunnerWithParameters did it a little bit differently,
because it was not possible to provide additional information for
validation via constructor. I added a second constructor to
ParentRunner and BlockJUnit4ClassRunner that makes it possible to
provide validators that are executed additionally to the current
validation methods.