Fix deprecation message of Assert#assertThat
The old deprecation message referenced the class
org.hamcrest.junit.MatcherAssert of the hamcrest-junit artifact.
hamcrest-junit was an experiment that is not continued. There is now a
Hamcrest 2.1 release that still contains the class
org.hamcrest.MatcherAssert like older releases e.g. Hamcrest 1.3. This
means that the deprecation message is now valid both for users using
Hamcrest 1.3 and users using Hamcrest 2.1.